fix: maintain consistent activeCard amount (prevent not enough entries at the end w/multiple-choice)
Instead of sampling from the first index, it moves the index of the displayed (current) card up to the 4th (out of 10) and continues to cycle cards the the end when they reach the 0 index. When it can't draw from the inactive pile (it's empty), it moves the index up until it reaches 10.

The initial offset of 4 is to prevent all 10 of the cards from being seen for all 10 (at least) times it takes to move from index 0 to 10 at the end when there are no more inactive cards. With an offset of 4, they're all only shown 6 times instead of 10. And at the beginning, it will sample the exact same set 4 times instead of 0 (splitting the difference here).